Split library into classes
* Rule
* Modifier
* ValidationException
