+++ src/Modifier.js
@@ -1,0 +1,9 @@
+//  TODO: add docs
+class Modifier {
+  constructor(fork, exec) {
+    this.fork = fork;
+    this.exec = exec;
+  }
+}
+
+export default Modifier;
