+++ src/v8n.js
@@ -1,3 +1,7 @@
+import Rule from "./Rule";
+import Modifier from "./Modifier";
+import ValidationException from "./ValidationException";
+
 /**
  * Function used to produce a {@link Validation} object. The Validation object
  * is used to configure a validation strategy and perform the validation tests.
@@ -179,81 +183,6 @@ function applyRule(rule, name) {
 }
 
 /**
- * A Rule object instance stores information about a rule inside the validation
- * process.
- *
- * > It's mostly used by the developer to handle validation results. It's
- * > instantiated automatically by the library engine during the validation
- * > process and this should not be done directly by the developer.
- */
-class Rule {
-  /**
-   * Constructs a Rule object instance.
-   *
-   * @param {string} name rule name
-   * @param {function} fn rule function used to perform the validation
-   * @param {*} args arguments used by the rule
-   * @param {Array} modifiers list of modifiers to be applied on the Rule
-   */
-  constructor(name, fn, args, modifiers) {
-    this.name = name;
-    this.fn = fn;
-    this.args = args;
-    this.modifiers = modifiers;
-  }
-
-  _test(value) {
-    const modifiers = this.modifiers.slice();
-    const fn = this.fn;
-
-    if (modifiers.length) {
-      const first = modifiers.pop();
-      value = first.fork(fn, value);
-      value = first.exec(value);
-
-      while (modifiers.length) {
-        const modifier = modifiers.pop();
-        value = modifier.fork(it => it, value);
-        value = modifier.exec(value);
-      }
-      return value;
-    } else {
-      return fn(value);
-    }
-  }
-
-  _testAsync(value) {
-    const modifiers = this.modifiers.slice();
-    const fn = this.fn;
-
-    if (modifiers.length) {
-      const first = modifiers.pop();
-      value = first.fork(val => Promise.resolve(fn(val)), value);
-
-      const isArray = Array.isArray(value);
-
-      return Promise.all(isArray ? value : [value])
-        .then(result => {
-          let value = first.exec(isArray ? result : result[0]);
-
-          while (modifiers.length) {
-            const modifier = modifiers.pop();
-            value = modifier.fork(it => it, value);
-            value = modifier.exec(value);
-          }
-
-          return value;
-        })
-        .catch(ex => {
-          return false;
-        });
-    } else {
-      return Promise.resolve(fn(value));
-    }
-  }
-}
-
-/**
  * Group of functionalities that can be performed on a validation object.
  *
  * > This object should not be used directly. All of its functionalities will be
@@ -379,48 +308,6 @@ function executeAsyncRulesAux(value, rules, resolve, reject) {
     }
   } else {
     resolve(value);
-  }
-}
-
-/**
- * Exception which represents a validation issue.
- *
- * It contains information about the {@link Rule} which was being performed when
- * the issue happened, and about the value which was being validated.
- *
- * > An exception object can be used as a chain for handling nested validation
- * > results. If some validation is composed by other validations, the `cause`
- * > property of the exception can be used to get the next deepest level in the
- * > error chain.
- */
-class ValidationException extends Error {
-  /**
-   * Constructs a validation exception with the rule which caused the issue and
-   * the value which was being validated when the issue happened.
-   *
-   * @param {Rule} rule the rule object which caused the validation
-   * @param {any} value the validated value
-   * @param {Error} cause indicates which problem ocurred during the validation;
-   * it can be used as chain to detected deep validations
-   * @param {string} target? indicates the target which was being validated, it
-   * can be a key in a object validation, for example
-   */
-  constructor(rule, value, cause, target, ...remaining) {
-    super(remaining);
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, ValidationException);
-    }
-    this.rule = rule;
-    this.value = value;
-    this.cause = cause;
-    this.target = target;
-  }
-}
-
-class Modifier {
-  constructor(fork, exec) {
-    this.fork = fork;
-    this.exec = exec;
   }
 }
